app/main.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import List
import pandas as pd
from datetime import datetime
import os
import logging

app = FastAPI()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class StockData(BaseModel):
    date: str
    close: float
    volume: int

@app.get("/api/v1/stock_data", response_model=List[StockData])
async def get_stock_data(symbol: str, start_date: str, end_date: str):
    """
    Fetch stock data for a given symbol and date range.
    Placeholder implementation; replace with actual data source (e.g., PostgreSQL, Kafka).
    """
    try:
        # Validate inputs
        start = datetime.strptime(start_date, '%Y-%m-%d')
        end = datetime.strptime(end_date, '%Y-%m-%d')
        if start >= end:
            raise ValueError("Start date must be before end date")
        
        # Placeholder: Generate sample data
        # In production, fetch from PostgreSQL or Kafka
        dates = pd.date_range(start=start, end=end, freq='D')
        data = [
            {
                "date": date.strftime('%Y-%m-%d'),
                "close": 100 + i * 0.5,  # Sample close price
                "volume": 1000 + i * 100  # Sample volume
            }
            for i, date in enumerate(dates)
        ]
        
        logger.info(f"Returning {len(data)} records for {symbol}")
        return data
    
    except ValueError as e:
        logger.error(f"Invalid input: {str(e)}")
        raise HTTPException(status_code=400, detail=str(e))
    except Exception as e:
        logger.error(f"Error fetching data: {str(e)}")
        raise HTTPException(status_code=500, detail="Internal server error")

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)